Implement basic coloured square detection
- Add frame skipper. This can be used to skip some frames if they take
  too long to process, since not all frames need to be processed.
- Execute OpenCV stuff on a new thread. This makes the video less laggy
  and (at this point) eliminates the need for a frame skipper.

The coloured squares are detected based on their RGB values. Any pixel
within the colour range for a certain colour is assumed to be in the
square. A better method will probably need to be used. A border is drawn
around the detected squares with the colour written in the middle.